Rename `worker.rs` to `system_domain_worker.rs`
No logical changes
