+++ domains/client/domain-executor/src/system_bundle_producer.rs
@@ -1,5 +1,5 @@
 use crate::bundle_election_solver::BundleElectionSolver;
-use crate::worker::ExecutorSlotInfo;
+use crate::system_domain_worker::ExecutorSlotInfo;
 use crate::{BundleSender, ExecutionReceiptFor};
 use codec::{Decode, Encode};
 use futures::{select, FutureExt};
